refactor: switch to fs/promises + try-catch approach
