+++ src/fs/delete.js
@@ -1,13 +1,14 @@
-import { unlink } from 'fs';
+import { unlink } from 'fs/promises';
 const PROJ_DIR = process.cwd();
 const FS_ERROR = 'FS operation failed';
 
 const remove = async () => {
-    await unlink(PROJ_DIR + '/src/fs/files/fileToRemove.txt', (err) => {
-        if (err) {
-            throw new Error(FS_ERROR);
-        }
-    });
+    try {
+        await unlink(PROJ_DIR + '/src/fs/files/fileToRemove.txt');
+    }
+    catch {
+        throw new Error(FS_ERROR);
+    }
 };
 
 await remove();
