+++ src/fs/read.js
@@ -1,14 +1,15 @@
-import { readFile } from 'fs';
+import { readFile } from 'fs/promises';
 const PROJ_DIR = process.cwd();
 const FS_ERROR = 'FS operation failed';
 
 const read = async () => {
-    await readFile(PROJ_DIR + '/src/fs/files/fileToRead.txt', 'utf8', (err, data) => {
-        if (err) {
-            throw new Error(FS_ERROR);
-        }
+    try {
+        const data = await readFile(PROJ_DIR + '/src/fs/files/fileToRead.txt', 'utf8');
         console.log(data);
-    });
+    }
+    catch {
+        throw new Error(FS_ERROR);
+    }
 };
 
 await read();
