+++ src/fs/rename.js
@@ -1,4 +1,5 @@
-import { access, rename as fsrename, constants } from 'fs';
+import { access, constants } from 'fs';
+import { rename as fsrename } from 'fs/promises';
 const PROJ_DIR = process.cwd();
 const FS_ERROR = 'FS operation failed';
 
@@ -9,11 +10,12 @@ const rename = async () => {
         }
     });
 
-    await fsrename(PROJ_DIR + '/src/fs/files/wrongFilename.txt', PROJ_DIR + '/src/fs/files/properFilename.md', (err) => {
-        if (err) {
-            throw new Error(FS_ERROR);
-        }
-    });
+    try {
+        await fsrename(PROJ_DIR + '/src/fs/files/wrongFilename.txt', PROJ_DIR + '/src/fs/files/properFilename.md');
+    }
+    catch {
+        throw new Error(FS_ERROR);
+    }
 };
 
 await rename();
